listener/requirements.txt
telethon>=1.36
python-dotenv>=1.0
aiokafka>=0.11
